build-epiphany-sdk.sh: Symlink epiphany-elf in host to toolchain directory
The ESDK directory structure is split into tools/e-gnu.$arch and
tools/host.$arch.

PAL is configured to install the Epiphany device libraries/headers in
tools/host.$arch/epiphany-elf/{lib,include}, but the default search path
for the epiphany toolchain is in tools/e-gnu.$arch/epiphany-elf.
Adding the symlink is the easiest way to make the files end up in the
right path.

/
	* build-epiphany-sdk.sh: Symlink epiphany-elf in host to
	epiphany-elf in e-gnu toolchain.

Signed-off-by: Ola Jeppsson <ola@adapteva.com>